First steps towards a better application structure with blueprints
