+++ README.md
@@ -1,6 +1,8 @@
 there is a new environment variable called: DATABASE_URL
 through that variable you can configure your own database to run the app with
 
+Also: REDIS_URL can be used
+
 <img src="images/normalLogo.png" alt="MusicServiceLogo" align="left" width="130"/>
 
 # Music sync Service (youtube-dl-sync) <br> ![Release version](https://img.shields.io/github/v/release/thijstakken/musicservice?label=latest) ![Docker Pulls](https://img.shields.io/docker/pulls/thijstakken/musicservice?label=downloads) ![DevOps Build](https://img.shields.io/azure-devops/build/mydevCloudThijsHVA/d94ee522-5f5b-43cf-a9e3-175c5cf1fb03/3) ![License](https://img.shields.io/github/license/thijstakken/musicservice) ![Issues](https://img.shields.io/github/issues/thijstakken/musicservice)
