+++ webapp/routes.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals
 from datetime import datetime
 from flask import Flask, render_template, request, redirect, url_for, send_file, flash, Blueprint
-from flask_login import current_user, login_user, logout_user, login_required
+from flask_login import current_user, login_required
 import sqlalchemy as sa
 from webapp import db
 from webapp.models import User
@@ -21,13 +21,7 @@ import os.path
 import time
 import sys
 from pathlib import Path
-import schedule
 import threading
-
-# import the downloadmusic function from the downloadMusic.py file
-#from downloadMusic import downloadmusic
-#from uploadMusic import uploadmusic
-from webapp.downloadScheduler import scheduleJobs, deleteJobs, immediateJob, run_schedule
 
 from webapp import app
 
@@ -46,211 +40,6 @@ from sqlalchemy import select
 @login_required
 def profile():
     return render_template('profile.html')
-
-@app.route("/", methods=["GET", "POST"])
-@login_required
-def musicapp():
-    # get the music list from the database with scalars
-    #music_list = db.session.scalars(sa.select(Music)).all()
-
-    # get the music_list but only for the logged in user
-    music_list = db.session.scalars(sa.select(Music).where(Music.user_id == current_user.id)).all()
-
-    form = MusicForm()
-    if form.validate_on_submit():
-        music = Music()
-        music.user_id = current_user.id
-        music.title = form.title.data
-        music.url = form.url.data
-        music.monitored = form.monitored.data
-        music.interval = form.interval.data
-        db.session.add(music)
-        db.session.commit()
-        flash('Song added')
-        return redirect(url_for('musicapp'))
-
-    return render_template("musicapp.html", music_list=music_list, form=form)
-    #return "musicapppage"
-
-@app.route("/add", methods=["POST"])
-@login_required
-def add():
-    title = request.form.get("title")
-    url = request.form.get("url")
-    new_music = Music()
-    new_music.title = title
-    new_music.url = url
-    new_music.monitored = False
-    new_music.interval = 10
-    db.session.add(new_music)
-    db.session.commit()
-    
-    # at the moment, the schedule is always false upon creation. so this is not needed at the moment
-    # this has already been used in the monitor function below this function.
-    #if new_music.monitored is False:
-        # schedule a job for the newly added playlist/song with the corrosponding interval value
-    #    scheduleJobs(music_id, title, interval)
-    return redirect(url_for("musicapp"))
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('musicapp'))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = db.session.scalar(
-            sa.select(User).where(User.username == form.username.data))
-        if user is None or not user.check_password(form.password.data):
-            flash('Invalid username or password')
-            return redirect(url_for('login'))
-        login_user(user, remember=form.remember_me.data)
-        next_page = request.args.get('next')
-        if not next_page or urlsplit(next_page).netloc != '':
-            next_page = url_for('musicapp')
-        return redirect(next_page)
-    return render_template('login.html', title='Sign In', form=form)
-
-@app.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for('musicapp'))
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('musicapp'))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        user = User(username=form.username.data)
-        user.set_password(form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        flash('Congratulations, you are now a registered user!')
-        return redirect(url_for('login'))
-    return render_template('register.html', title='Register', form=form)
-
-@app.route("/monitor/<int:music_id>")
-@login_required
-def monitor(music_id):
-    # get the music object from the database with scalars
-    music = db.session.scalars(sa.select(Music).where(Music.id == music_id)).first()
-
-    # turned below rule off because during startup the settings are already set.
-    #settings = WebDAV.query.filter_by(id=1).first()
-    if music is not None:
-        music.monitored = not music.monitored
-        db.session.commit()
-        if music.monitored is True and settings is not None:
-            print("monitor is ON")
-            print("Going to schedule the music to be downloaded on repeat")
-            print(music.monitored)
-            # schedule a job for the newly added playlist/song with the corrosponding interval value
-            scheduleJobs(music, settings)
-            # add flash message to confirm the interval change
-            flash('Monitoring: On for ' + str(music.title))
-        elif music.monitored is False:
-            print("monitor is OFF")
-            print("Going to delete the scheduled job")
-            #print(music.monitored)
-            # delete the scheduled job for the deleted playlist/song
-            deleteJobs(music.id)
-            # add flash message to confirm the interval change
-            flash('Monitoring: Off for ' + str(music.title))
-    return redirect(url_for("musicapp"))
-
-@app.route("/delete/<int:music_id>")
-@login_required
-def delete(music_id):
-
-    # get the music object from the database with scalars
-    music = db.session.scalars(sa.select(Music).where(Music.id == music_id)).first()
-
-    # check if song exists
-    if music is None:
-        flash('Song not found')
-        return redirect(url_for("musicapp"))
-    
-    # check if the user is the owner of the song
-    #print("music user id:", music.user_id)
-    #print("current user id:", current_user.id)
-    if music.user_id != current_user.id:
-        flash('You cannot delete songs of others!')
-        return redirect(url_for("musicapp"))
-
-    # delete the music object from the database
-    db.session.delete(music)
-    db.session.commit()
-    # delete the scheduled job for the deleted playlist/song
-    deleteJobs(music_id)
-    return redirect(url_for("musicapp"))
-
-@app.route("/download/<int:music_id>")
-@login_required
-def download(music_id):
-    # get the music object from the database with scalars
-    music = db.session.scalars(sa.select(Music).where(Music.id == music_id)).first()
-    # execute the download function to download one time
-    if music is not None and settings is not None:
-        immediateJob(music, settings)
-    return redirect(url_for("musicapp"))
-
-# let users configure their interval value on a per playlist/song basis
-@app.route("/interval/<int:music_id>")
-@login_required
-def interval(music_id):
-    
-    # at the moment it accepts everything. but it should only allow integers as input.
-    # close this down somewhere so only integers are allowed through this method.
-    interval = request.args.get('interval', None) # None is the default value if no interval is specified
-    
-    # get the music object from the database with scalars
-    music = db.session.scalars(sa.select(Music).where(Music.id == music_id)).first()
-
-
-    # get the CloudStorage settings from the database with scalars
-    settings = db.session.scalars(sa.select(CloudStorage).where(CloudStorage.id == music_id)).first()
-
-    # settings = WebDAV.query.filter_by(id=1).first()
-
-
-    if music:
-        music.interval = interval
-        db.session.commit()
-        #print(interval)
-
-        # add flash message to confirm the interval change
-        flash('Interval changed to ' + str(interval) + ' minutes')
-        
-        # if the monitor is on, then reschedule the job with the new interval value
-        if music.monitored is True:
-            print("Going to reschedule the music to be downloaded on repeat")
-            # delete the scheduled job for the deleted playlist/song
-            deleteJobs(music_id)
-            # schedule a job for the newly added playlist/song with the corrosponding interval value
-            scheduleJobs(music, settings)
-
-    return redirect(url_for("musicapp"))
-
-# this function can get the time left before the playlist will be downloaded again
-@app.route("/intervalstatus/<int:music_id>")
-@login_required
-def intervalStatus(music_id):
-    
-    time_of_next_run = schedule.next_run(music_id)
-    # get current time
-    time_now = datetime.now()
-    
-    if time_of_next_run is not None:
-        # calculate time left before next run
-        time_left = time_of_next_run - time_now
-        print("Time left before next run:", time_left)
-        time_left = time_left.seconds
-    else:
-        time_left = 0
-    
-    # return the time left before the next run
-    return str(time_left)
-
 
 
 ### START OF SETTINGS ###
