+++ webapp/templates/errors/404.html
@@ -1,0 +1,1 @@
+404 error template via blueprint
