+++ webapp/templates/errors/500.html
@@ -1,0 +1,1 @@
+500 error template via blueprint
